Feature/full fledged engine conversion (#42)
* Fix perft coverage and use parallel CMake build

* Phase 2: Setup ONNX directory, point train.py to it, and link ONNX Runtime via CMake

* Rename ONNX file to fenrir.onnx and fix coverage

* Implement ONNX hot-reloading with full test coverage

* Fix flaky opponent checkmate test

* Fix MSVC compiler flags

* Fix all MSVC compilation warnings and errors

* Fix C4101 in logger.cpp while preserving gcov optimizations

* Add torch-directml support to trainer for AMD GPUs on Windows

* Remove test break in training loop to run continuously

* Add script to launch multiple self-play instances on Windows

* Fix crashes caused by ONNX evaluation errors, and implement jsonl replay buffer cleanup

* Fix bug

* Fix MSVC C4101 unreferenced variable warning

* Fix uncaught exceptions in thread evaluator during hot-swap

* Fix file locking segfault by reading ONNX model into memory

* Pause python training loop when there is no new data

* Keep ONNX model memory buffer alive across session lifetime

* Safely catch thread creation exceptions in MCTS

* Fix runtime issues for training

* Fix ONNX file reload crashes, unhandled std::future_error, and PermissionErrors

* Fix fen_to_tensor bug and ONNX reload OOM crash

* Enable CUDA execution provider in ONNX Runtime

* Switch to DirectML for AMD/Intel GPU support

* Disable MemPattern for DirectML to enforce GPU execution

* Fix NuGet package extraction for DirectML by specifying DOWNLOAD_NAME

* Fix Windows min/max macro conflict

* Fix train.py massive overfitting on small dataset updates

* Fix PyTorch multiprocessing ancdata bug

* Apply massive PyTorch performance optimizations

* Remove torch.compile

* Enable FP16 Mixed Precision and TF32 Tensor Cores

* Reduce epochs per chunk to 1 to cut training time

* Add tests for UCI time management to reach 100% coverage

* Remove basic scripts

* Remove basic scripts

---------

+++ src/chess/fen.cpp
@@ -75,7 +75,7 @@ namespace fenrir
 	void Fen::validate_chess_rules(const std::string &placement_string) const
 	{
 		std::unordered_map<char, uint8_t> piece_counts = {
-			{'K', 0}, {'k', 0}, {'Q', 0}, {'q', 0}, {'R', 0}, {'r', 0}, {'B', 0}, {'b', 0}, {'N', 0}, {'n', 0}, {'P', 0}, {'p', 0}};
+			{'K', static_cast<uint8_t>(0)}, {'k', static_cast<uint8_t>(0)}, {'Q', static_cast<uint8_t>(0)}, {'q', static_cast<uint8_t>(0)}, {'R', static_cast<uint8_t>(0)}, {'r', static_cast<uint8_t>(0)}, {'B', static_cast<uint8_t>(0)}, {'b', static_cast<uint8_t>(0)}, {'N', static_cast<uint8_t>(0)}, {'n', static_cast<uint8_t>(0)}, {'P', static_cast<uint8_t>(0)}, {'p', static_cast<uint8_t>(0)}};
 
 		for (char c : placement_string)
 		{
